+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [v1.2.0] - 2021-10-15
 ### Added
 - Added the Icelandic NER dataset MIM-GOLD-NER. This can now be loaded as
   `mim-gold-ner` in the `Benchmark` class and through the CLI.